server/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-cors==0.8.1
aiosignal==1.4.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
docker==7.1.0
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
huggingface-hub==0.33.4
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
llava==0.0.1.dev0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
safetensors==0.5.3
scipy==1.16.0
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.2
stack-data==0.6.3
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
yarl==1.20.1
